Move `cip-mm` to `kpromo mm`
