+++ cmd/kpromo/cmd/mm/mm.go
@@ -14,101 +14,82 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package mm
 
 import (
 	"context"
-	"fmt"
 
-	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
 	reg "sigs.k8s.io/promo-tools/v3/legacy/dockerregistry"
-	"sigs.k8s.io/release-utils/log"
 )
 
-var cmd = &cobra.Command{
-	Short: "cip-mm → Container Image Promoter - Manifest Modificator",
-	Long: `cip-mm → Container Image Promoter - Manifest Modificator
+// TODO(cip-mm): Remove in the next minor release.
+
+var MMCmd = &cobra.Command{
+	Short: "[DEPRECATED] mm → Manifest Modifier",
+	Long: `[DEPRECATED] mm → Manifest Modifier
 
 This tool **m**odifies promoter **m**anifests. For now it dumps some filtered
 subset of a staging GCR and merges those contents back into a given promoter
 manifest.`,
-	Use:               "cip-mm",
-	SilenceUsage:      true,
-	SilenceErrors:     true,
-	PersistentPreRunE: initLogging,
+	Use:           "mm",
+	SilenceUsage:  true,
+	SilenceErrors: true,
 	RunE: func(cmd *cobra.Command, args []string) error {
 		return run()
 	},
 }
 
-type commandLineOptions struct {
+type modifyOptions struct {
 	baseDir      string
 	stagingRepo  string
 	filterImage  string
 	filterDigest string
 	filterTag    string
-	logLevel     string
 }
 
-var commandLineOpts = &commandLineOptions{}
+var modifyOpts = &modifyOptions{}
 
 func init() {
-	cmd.PersistentFlags().StringVar(
-		&commandLineOpts.baseDir,
+	MMCmd.PersistentFlags().StringVar(
+		&modifyOpts.baseDir,
 		"base_dir",
 		"",
 		"the manifest directory to look at and modify",
 	)
-	cmd.PersistentFlags().StringVar(
-		&commandLineOpts.stagingRepo,
+	MMCmd.PersistentFlags().StringVar(
+		&modifyOpts.stagingRepo,
 		"staging_repo",
 		"",
 		"the staging repo which we want to read from",
 	)
-	cmd.PersistentFlags().StringVar(
-		&commandLineOpts.filterImage,
+	MMCmd.PersistentFlags().StringVar(
+		&modifyOpts.filterImage,
 		"filter_image",
 		"",
 		"filter staging repo by this image name",
 	)
-	cmd.PersistentFlags().StringVar(
-		&commandLineOpts.filterDigest,
+	MMCmd.PersistentFlags().StringVar(
+		&modifyOpts.filterDigest,
 		"filter_digest",
 		"",
 		"filter images by this digest",
 	)
-	cmd.PersistentFlags().StringVar(
-		&commandLineOpts.filterTag,
+	MMCmd.PersistentFlags().StringVar(
+		&modifyOpts.filterTag,
 		"filter_tag",
 		"",
 		"filter images by this tag",
 	)
-	cmd.PersistentFlags().StringVar(
-		&commandLineOpts.logLevel,
-		"log-level",
-		"info",
-		fmt.Sprintf("the logging verbosity, either %s", log.LevelNames()),
-	)
-}
-
-func main() {
-	if err := cmd.Execute(); err != nil {
-		logrus.Fatal(err)
-	}
-}
-
-func initLogging(*cobra.Command, []string) error {
-	return log.SetupGlobalLogger(commandLineOpts.logLevel)
 }
 
 func run() error {
 	opt := reg.GrowManifestOptions{}
 	if err := opt.Populate(
-		commandLineOpts.baseDir, commandLineOpts.stagingRepo,
-		commandLineOpts.filterImage, commandLineOpts.filterDigest,
-		commandLineOpts.filterTag); err != nil {
+		modifyOpts.baseDir, modifyOpts.stagingRepo,
+		modifyOpts.filterImage, modifyOpts.filterDigest,
+		modifyOpts.filterTag); err != nil {
 		return err
 	}
