+++ cmd/kpromo/cmd/root.go
@@ -25,6 +25,7 @@ import (
 	"sigs.k8s.io/promo-tools/v3/cmd/kpromo/cmd/cip"
 	"sigs.k8s.io/promo-tools/v3/cmd/kpromo/cmd/gh"
 	"sigs.k8s.io/promo-tools/v3/cmd/kpromo/cmd/manifest"
+	"sigs.k8s.io/promo-tools/v3/cmd/kpromo/cmd/mm"
 	"sigs.k8s.io/promo-tools/v3/cmd/kpromo/cmd/pr"
 	"sigs.k8s.io/promo-tools/v3/cmd/kpromo/cmd/run"
 	"sigs.k8s.io/promo-tools/v3/cmd/kpromo/cmd/version"
@@ -80,6 +81,8 @@ func init() {
 	rootCmd.AddCommand(manifest.ManifestCmd)
 	rootCmd.AddCommand(pr.PRCmd)
 	rootCmd.AddCommand(version.VersionCmd)
+	// TODO(cip-mm): Remove in the next minor release.
+	rootCmd.AddCommand(mm.MMCmd)
 }
 
 func initLogging(*cobra.Command, []string) error {
